sell/src/test/java/com/imooc/sell/serviceImp/CategoryServiceImpTest.java
package com.imooc.sell.serviceImp;

import com.imooc.sell.dataobject.ProductCategory;
import org.junit.Assert;
import org.junit.Test;
import org.junit.runner.RunWith;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.context.SpringBootTest;
import org.springframework.test.context.junit4.SpringRunner;

import javax.swing.*;

import java.util.Arrays;
import java.util.List;

import static org.junit.Assert.*;
@RunWith(SpringRunner.class)
@SpringBootTest
public class CategoryServiceImpTest {
    @Autowired
    private  CategoryServiceImp categoryServiceImp;

    @Test
    public void findOne() throws  Exception{
        ProductCategory productCategory = categoryServiceImp.findOne(1);
        Assert.assertEquals(new Integer(1),productCategory.getCategoryId());
    }

    @Test
    public void findAll() throws  Exception{
        List<ProductCategory> productCategoryList = categoryServiceImp.findAll();
        Assert.assertNotEquals(0,productCategoryList.size());
    }

    @Test
    public void findByCategoryTypeIn() {
        List<ProductCategory> productCategoryList = categoryServiceImp.findByCategoryTypeIn(Arrays.asList(1,2,3,4,5));
        Assert.assertNotEquals(0,productCategoryList.size());
    }

    @Test
    public void save() {
        ProductCategory productCategory = new ProductCategory("男生专享",10);
        ProductCategory result = categoryServiceImp.save(productCategory);
        Assert.assertNotNull(productCategory);
    }
}